token is 16 bits now

+++ ArduCopter/GCS_Mavlink.pde
@@ -1617,7 +1617,7 @@ GCS_MAVLINK::_count_parameters()
 	// if we haven't cached the parameter count yet...
 	if (0 == _parameter_count) {
         AP_Param  *vp;
-        uint32_t token;
+        uint16_t token;
 
         vp = AP_Param::first(&token, NULL);
 		do {
